Create table.c
Place this file in usr/src/servers/pm . 
Replace the existing table.c is what I mean. 

Note: The numbers mentioned in this file should match the numbers mentioned in the calls in systemcalls.h. Like for example, here is 45 is IGLookUp, then even in systemcalls.h , the number should be 45 for IGLookUp